Unificación rama implementación UI radar
Implementación de la rama donde se desarrollo la UI de la actividad del radar
