+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">Multisalto</string>
-    <string name="multisaltos">Multisaltos</string>
+    <string name="multisaltos">Multisalto</string>
     <string name="Radar">Radar</string>
     <string name="welcome">"Bienvenido a Multisaltos"</string>
     <string name="resgister">"REGISTRARSE"</string>
